feat: security improve
