+++ field/field.go
@@ -15,35 +15,35 @@ type ScanValuer interface {
 type Field struct{ expr }
 
 func (field Field) Eq(value ScanValuer) Expr {
-	return expr{expression: clause.Eq{Column: field.RawExpr(), Value: value}}
+	return expr{e: clause.Eq{Column: field.RawExpr(), Value: value}}
 }
 
 func (field Field) Neq(value ScanValuer) Expr {
-	return expr{expression: clause.Neq{Column: field.RawExpr(), Value: value}}
+	return expr{e: clause.Neq{Column: field.RawExpr(), Value: value}}
 }
 
 func (field Field) In(values ...ScanValuer) Expr {
-	return expr{expression: clause.IN{Column: field.RawExpr(), Values: field.toSlice(values...)}}
+	return expr{e: clause.IN{Column: field.RawExpr(), Values: field.toSlice(values...)}}
 }
 
 func (field Field) Gt(value ScanValuer) Expr {
-	return expr{expression: clause.Gt{Column: field.RawExpr(), Value: value}}
+	return expr{e: clause.Gt{Column: field.RawExpr(), Value: value}}
 }
 
 func (field Field) Gte(value ScanValuer) Expr {
-	return expr{expression: clause.Gte{Column: field.RawExpr(), Value: value}}
+	return expr{e: clause.Gte{Column: field.RawExpr(), Value: value}}
 }
 
 func (field Field) Lt(value ScanValuer) Expr {
-	return expr{expression: clause.Lt{Column: field.RawExpr(), Value: value}}
+	return expr{e: clause.Lt{Column: field.RawExpr(), Value: value}}
 }
 
 func (field Field) Lte(value ScanValuer) Expr {
-	return expr{expression: clause.Lte{Column: field.RawExpr(), Value: value}}
+	return expr{e: clause.Lte{Column: field.RawExpr(), Value: value}}
 }
 
 func (field Field) Like(value ScanValuer) Expr {
-	return expr{expression: clause.Like{Column: field.RawExpr(), Value: value}}
+	return expr{e: clause.Like{Column: field.RawExpr(), Value: value}}
 }
 
 func (field Field) toSlice(values ...ScanValuer) []interface{} {
