+++ field/float.go
@@ -5,35 +5,35 @@ import "gorm.io/gorm/clause"
 type Float64 Field
 
 func (field Float64) Eq(value float64) Expr {
-	return expr{expression: clause.Eq{Column: field.RawExpr(), Value: value}}
+	return expr{e: clause.Eq{Column: field.RawExpr(), Value: value}}
 }
 
 func (field Float64) Neq(value float64) Expr {
-	return expr{expression: clause.Neq{Column: field.RawExpr(), Value: value}}
+	return expr{e: clause.Neq{Column: field.RawExpr(), Value: value}}
 }
 
 func (field Float64) Gt(value float64) Expr {
-	return expr{expression: clause.Gt{Column: field.RawExpr(), Value: value}}
+	return expr{e: clause.Gt{Column: field.RawExpr(), Value: value}}
 }
 
 func (field Float64) Gte(value float64) Expr {
-	return expr{expression: clause.Gte{Column: field.RawExpr(), Value: value}}
+	return expr{e: clause.Gte{Column: field.RawExpr(), Value: value}}
 }
 
 func (field Float64) Lt(value float64) Expr {
-	return expr{expression: clause.Lt{Column: field.RawExpr(), Value: value}}
+	return expr{e: clause.Lt{Column: field.RawExpr(), Value: value}}
 }
 
 func (field Float64) Lte(value float64) Expr {
-	return expr{expression: clause.Lte{Column: field.RawExpr(), Value: value}}
+	return expr{e: clause.Lte{Column: field.RawExpr(), Value: value}}
 }
 
 func (field Float64) In(values ...float64) Expr {
-	return expr{expression: clause.IN{Column: field.RawExpr(), Values: field.toSlice(values...)}}
+	return expr{e: clause.IN{Column: field.RawExpr(), Values: field.toSlice(values...)}}
 }
 
 func (field Float64) NotIn(values ...float64) Expr {
-	return expr{expression: clause.Not(field.In(values...))}
+	return expr{e: clause.Not(field.In(values...).expression())}
 }
 
 func (field Float64) Between(left float64, right float64) Expr {
@@ -45,11 +45,11 @@ func (field Float64) NotBetween(left float64, right float64) Expr {
 }
 
 func (field Float64) Like(value float64) Expr {
-	return expr{expression: clause.Like{Column: field.RawExpr(), Value: value}}
+	return expr{e: clause.Like{Column: field.RawExpr(), Value: value}}
 }
 
 func (field Float64) NotLike(value float64) Expr {
-	return expr{expression: clause.Not(field.Like(value))}
+	return expr{e: clause.Not(field.Like(value).expression())}
 }
 
 func (field Float64) Add(value float64) Float64 {
@@ -87,35 +87,35 @@ func (field Float64) toSlice(values ...float64) []interface{} {
 type Float32 Float64
 
 func (field Float32) Eq(value float32) Expr {
-	return expr{expression: clause.Eq{Column: field.RawExpr(), Value: value}}
+	return expr{e: clause.Eq{Column: field.RawExpr(), Value: value}}
 }
 
 func (field Float32) Neq(value float32) Expr {
-	return expr{expression: clause.Neq{Column: field.RawExpr(), Value: value}}
+	return expr{e: clause.Neq{Column: field.RawExpr(), Value: value}}
 }
 
 func (field Float32) Gt(value float32) Expr {
-	return expr{expression: clause.Gt{Column: field.RawExpr(), Value: value}}
+	return expr{e: clause.Gt{Column: field.RawExpr(), Value: value}}
 }
 
 func (field Float32) Gte(value float32) Expr {
-	return expr{expression: clause.Gte{Column: field.RawExpr(), Value: value}}
+	return expr{e: clause.Gte{Column: field.RawExpr(), Value: value}}
 }
 
 func (field Float32) Lt(value float32) Expr {
-	return expr{expression: clause.Lt{Column: field.RawExpr(), Value: value}}
+	return expr{e: clause.Lt{Column: field.RawExpr(), Value: value}}
 }
 
 func (field Float32) Lte(value float32) Expr {
-	return expr{expression: clause.Lte{Column: field.RawExpr(), Value: value}}
+	return expr{e: clause.Lte{Column: field.RawExpr(), Value: value}}
 }
 
 func (field Float32) In(values ...float32) Expr {
-	return expr{expression: clause.IN{Column: field.RawExpr(), Values: field.toSlice(values...)}}
+	return expr{e: clause.IN{Column: field.RawExpr(), Values: field.toSlice(values...)}}
 }
 
 func (field Float32) NotIn(values ...float32) Expr {
-	return expr{expression: clause.Not(field.In(values...))}
+	return expr{e: clause.Not(field.In(values...).expression())}
 }
 
 func (field Float32) Between(left float32, right float32) Expr {
@@ -127,11 +127,11 @@ func (field Float32) NotBetween(left float32, right float32) Expr {
 }
 
 func (field Float32) Like(value float32) Expr {
-	return expr{expression: clause.Like{Column: field.RawExpr(), Value: value}}
+	return expr{e: clause.Like{Column: field.RawExpr(), Value: value}}
 }
 
 func (field Float32) NotLike(value float32) Expr {
-	return expr{expression: clause.Not(field.Like(value))}
+	return expr{e: clause.Not(field.Like(value).expression())}
 }
 
 func (field Float32) Add(value float32) Float32 {
